Material Fix (#139)
* Materials support multiple objects and Warning Fixes

Also fixed warnings in Animation with doubles and floats

* Global Illumination Fixed

+++ Source/YTE/Core/Actions/ActionManager.cpp
@@ -34,7 +34,7 @@ namespace YTE
     for (auto it = mSequences.begin(); it != mSequences.end(); ++it)
     {
       auto& sequence = it->second;
-      sequence.Increment(aUpdate->Dt);
+      sequence.Increment(static_cast<float>(aUpdate->Dt));
       sequence();
       if (sequence.IsDone())
       {
